Changes in buy offers
