+++ AI_NETCORE_API/AI_NETCORE_API/Models/Objects/BuyOfferModel.cs
@@ -10,7 +10,6 @@ namespace AI_NETCORE_API.Models.Objects
         public int Id { get; set; }
         public int Amount { get; set; }
         public int StartAmount { get; set; }
-        public int ResourceId { get; set; }
         public double MaxPrice { get; set; }
         public DateTime Date { get; set; }
         public bool IsValid { get; set; }
