+++ AI_NETCORE_API/AI_NETCORE_API/Startup.cs
@@ -64,6 +64,8 @@ using Domain.Creators.BuyOffer.Concrete;
 using Domain.Creators.SellOffer.Concrete;
 using Domain.Creators.SellOffer.Abstract;
 using Microsoft.EntityFrameworkCore;
+using Domain.Updaters.Buyoffers.Abstract;
+using Domain.Updaters.Buyoffers.Concrete;
 
 namespace AI_NETCORE_API
 {
@@ -105,7 +107,8 @@ namespace AI_NETCORE_API
             services.AddTransient<IBusinessObjectToModelsConverter, BusinessObjectToModelsConverter>();
             services.AddTransient<IBuyOffersProvider, BuyOffersProvider>();
             services.AddTransient<ISellOfferProvider, SellOfferProvider>();
-            services.AddTransient<IBuyOfferCreator, BuyOfferCreator>(); 
+            services.AddTransient<IBuyOfferCreator, BuyOfferCreator>();
+            services.AddTransient<IBuyOfferUpdater, BuyOfferUpdater>();
             services.AddTransient<IUserCreator, UserCreator>();
             services.AddTransient<ISellOfferCreator, SellOfferCreator>();
             services.AddTransient<IConfigurationUpdater, ConfigurationUpdater>();
