+++ AI_NETCORE_API/Domain/Providers/BuyOffers/Concrete/BuyOffersProvider.cs
@@ -38,19 +38,5 @@ namespace Domain.Providers.BuyOffers.Concrete
             }
         }
 
-        public IWithdrawBuyOfferByIdResponse WithdrawBuyOfferById(IWithdrawBuyOfferByIdRequest withdrawBuyOfferByIdRequest)
-        {
-            try
-            {
-                var result = _buyOffers.WithdrawBuyOffer(withdrawBuyOfferByIdRequest.BuyOfferId);
-                return new WithdrawBuyOfferByIdResponse(result);
-
-            }
-            catch (Exception ex)
-            {
-                _logger.Log(ex);
-                return new WithdrawBuyOfferByIdResponse();
-            }
-        }
     }
 }
