+++ AI_NETCORE_API/Domain/Updaters/BuyOffers/Abstract/IBuyOfferUpdater.cs
@@ -1,0 +1,13 @@
+﻿using Domain.Updaters.BuyOffers.Request.Abstract;
+using Domain.Updaters.BuyOffers.Response.Abstract;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Domain.Updaters.Buyoffers.Abstract
+{
+    public interface IBuyOfferUpdater
+    {
+        IWithdrawBuyOfferByIdResponse WithdrawBuyOfferById(IWithdrawBuyOfferByIdRequest withdrawBuyOfferByIdRequest);
+    }
+}
